+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/mobs/HoneySlimeEntity.java
@@ -165,6 +165,26 @@ public class HoneySlimeEntity extends AnimalEntity implements IAngerable, IMob {
    }
 
    @Override
+   public boolean onLivingFall(float distance, float damageMultiplier) {
+      if (distance > 1.0F) {
+         this.playSound(SoundEvents.BLOCK_SLIME_BLOCK_STEP, 0.4F, 1.0F);
+      }
+
+      int fallDamage = this.calculateFallDamage(distance, damageMultiplier);
+      if(this.isInHoney()){
+         fallDamage = (int)((fallDamage * 0.35f) - 3);
+      }
+
+      if (fallDamage <= 0) {
+         return false;
+      } else {
+         this.attackEntityFrom(DamageSource.FALL, (float)fallDamage);
+         this.playFallSound();
+         return true;
+      }
+   }
+
+   @Override
    public ActionResultType func_230254_b_(PlayerEntity player, Hand hand) {
       ItemStack itemstack = player.getHeldItem(hand);
       World world = player.getEntityWorld();